+++ cli/addmfadevice.go
@@ -9,6 +9,7 @@ import (
 	"github.com/aws/aws-sdk-go/aws/session"
 	"gopkg.in/alecthomas/kingpin.v2"
 
+	"github.com/99designs/aws-vault/mfa"
 	"github.com/99designs/aws-vault/vault"
 )
 
@@ -70,7 +71,7 @@ func AddMFADeviceCommand(app *kingpin.Application, input AddMfaDeviceCommandInpu
 
 	tokenProvider := input.Config.GetMfaTokenProvider()
 
-	im := vault.NewIAMMfa(sess, tokenProvider)
+	im := mfa.NewIAMMfa(sess, tokenProvider)
 
 	fmt.Printf("Adding MFA device to user %s using profile %s\n", input.Username, p.Name)
